requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
python-multipart
passlib[bcrypt]
uvicorn
python-dotenv
cloudinary
slowapi
email-validator
pytest 
pytest-cov 
httpx